+++ Wave/MenuParticleEntity.h
@@ -6,7 +6,7 @@ class EntityHandler;
 
 class MenuParticleEntity : public Entity { 
 public:
-	MenuParticleEntity(int32_t x, int32_t y, ID id, uint32_t uid, sf::Vector2f horizontalBounds, sf::Vector2f verticalBounds, EntityHandler& e, bool overGui, int32_t& scale);
+	MenuParticleEntity(int32_t x, int32_t y, ID id, uint32_t uid, sf::Vector2f horizontalBounds, sf::Vector2f verticalBounds, EntityHandler& e, bool overGui, double& scale);
 
 	virtual void render(sf::RenderWindow& w);
 	virtual void tick();
@@ -76,7 +76,7 @@ private:
 	sf::RectangleShape r;
 	sf::Color color;
 
-	int32_t& scale;
+	double& scale;
 	
 	sf::Vector2f horizontalBounds;
 	sf::Vector2f verticalBounds;
